+++ src/fflib/AFunction.hpp
@@ -2471,7 +2471,20 @@ struct evalE_F2 {
    }
 };
 
-//
+// Add FH juin 2020 
+template<typename A,typename B>
+class  OneBinaryOperatorBug : public OneOperator{
+    typedef  bool R;
+public:
+    E_F0 * code(const basicAC_F0 & args) const
+    {  //cout << "A op B \n" ;
+        cout << " Error expression impossible beetween "<< typeid(A).name() << " ,  "<<typeid(B).name() <<endl;
+        CompileError( "  Expression must return bad type!!!");
+        return  0;}
+    OneBinaryOperatorBug():
+    OneOperator(map_type[typeid(R).name()],map_type[typeid(A).name()],map_type[typeid(B).name()])
+    {pref = 100;}
+};
 template<typename C,class MI=OneBinaryOperatorMI,class MIx=evalE_F2 >
 class  OneBinaryOperator : public OneOperator{
   typedef  typename C::result_type R;